+++ packages/docs/src/stories/Tooltip.stories.tsx
@@ -1,15 +1,27 @@
 import React from 'react'
 import type { Meta, StoryObj } from '@storybook/react'
-import { Tooltip, TooltipProps, Box } from '@arturmp-ignite-ui/react'
-import * as RadixTooltip from '@radix-ui/react-tooltip'
-import { colors, fonts, radii, lineHeights } from '@arturmp-ignite-ui/tokens'
+import { Tooltip, TooltipProps, Box, Button } from '@arturmp-ignite-ui/react'
 
 export default {
   title: 'Data display/Tooltip',
   component: Tooltip,
   args: {
+    asChild: true,
     date: '10 de outubro',
     isAvailable: true,
+    sideOffset: -16,
+    children: (
+      <Button>
+        Hi there
+      </Button>
+    ),
+  },
+  argTypes: {
+    children: {
+      control: {
+        type: null
+      }
+    }
   },
   decorators: [
     (Story) => {
@@ -24,33 +36,7 @@ export default {
             gap: '$2'
           }}
         >
-          <RadixTooltip.Provider>
-            <RadixTooltip.Root>
-              <div style={{
-                width: 66,
-                height: 58,
-                borderRadius: radii.md,
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                backgroundColor: colors.gray600,
-              }}>
-                <RadixTooltip.Trigger asChild>
-                  <button style={{
-                    all: 'unset',
-                    color: colors.white,
-                    font: fonts.default,
-                    lineHeight: lineHeights.base,
-                  }}>
-                    10
-                  </button>
-                </RadixTooltip.Trigger>
-              </div>
-              <RadixTooltip.Portal>
-                {Story()}
-              </RadixTooltip.Portal>
-            </RadixTooltip.Root>
-          </RadixTooltip.Provider>
+          {Story()}
         </Box>
       )
     },