+++ packages/react/src/components/Tooltip/index.tsx
@@ -1,15 +1,28 @@
-import { TooltipArrow, TooltipContent } from "./styles";
+import * as RadixTooltip from '@radix-ui/react-tooltip'
+import { ComponentProps, ReactNode } from 'react';
+import { TooltipArrow, TooltipContent, TooltipTrigger } from "./styles";
 
-export interface TooltipProps {
+export interface TooltipProps extends ComponentProps<typeof RadixTooltip.Trigger> {
   date: string;
   isAvailable: boolean;
+  children: ReactNode;
+  sideOffset: number;
 }
 
-export function Tooltip({ date, isAvailable }: TooltipProps) {
+export function Tooltip({ date, isAvailable, children, sideOffset }: TooltipProps) {
   return (
-    <TooltipContent sideOffset={2}>
-      {date} - {isAvailable ? 'Disponível' : 'Indisponível'}
-      <TooltipArrow />
-    </TooltipContent>
+    <RadixTooltip.Provider>
+      <RadixTooltip.Root>
+        <TooltipTrigger>
+          {children}
+        </TooltipTrigger>
+        <RadixTooltip.Portal>
+        <TooltipContent sideOffset={sideOffset}>
+          {date} - {isAvailable ? 'Disponível' : 'Indisponível'}
+          <TooltipArrow />
+        </TooltipContent>
+        </RadixTooltip.Portal>
+      </RadixTooltip.Root>
+    </RadixTooltip.Provider>
   )
 }