+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ignite-ui/react
 
+## 4.0.0
+
+### Major Changes
+
+- Improving tooltip component
+
+## 3.0.0
+
+### Major Changes
+
+- Creating Tooltip and Toast components
+
 ## 2.0.0
 
 ### Major Changes